+++ zakura-network/src/peer_set.rs
@@ -13,4 +13,7 @@ pub(crate) use limit::{ActiveConnectionCounter, ConnectionTracker};
 use inventory_registry::InventoryRegistry;
 pub(crate) use set::PeerSet;
 
-pub use initialize::{init, init_with_zakura_header_sync};
+pub use initialize::{
+    init, init_with_zakura_header_sync,
+    init_with_zakura_header_sync_and_compatibility_pruning_height,
+};